+++ CHANGELOG.md
@@ -3,12 +3,13 @@ https://keepachangelog.com/en/1.0.0/
 
 ## [1.19.0] - 2023-01-11
 - `FormReadSynchronously`: Synchronously accessed `HttpRequest.Form` which uses sync-over-async. Use `HttpRequest.ReadFormAsync()` instead
+- `NewtonsoftMixedWithSystemTextJson` tests use `VerifyCS`
 
 ## [1.18.0] - 2023-01-10
 - `MultipleOrderByCalls`: Successive `OrderBy()` calls will maintain only the last specified sort order. Use `ThenBy()` to combine them
 - `TestMethodWithoutTestAttribute`: No longer fires for test constructors
 - `TestMethodWithoutTestAttribute`: No longer fires for overridden methods
-- `TestMethodWithoutTestAttribute` tests use VerifyCS
+- `TestMethodWithoutTestAttribute` tests use `VerifyCS`
 
 ## [1.17.6] - 2023-01-08
 - Switched several test files over to the `VerifyCS` approach